Ensure primary_size set before using

+++ src/wp-admin/includes/image.php
@@ -595,7 +595,7 @@ function _wp_make_subsizes( $new_sizes, $file, $image_meta, $attachment_id, $mim
 					$new_size     = isset( $new_size_meta['filesize'] ) ? $new_size_meta['filesize'] : 0;
 					$primary_size = isset( $image_meta['sizes'][ $new_size_name ]['filesize'] ) ? $image_meta['sizes'][ $new_size_name ]['filesize'] : 0;
 
-					if ( $new_size && $new_size >= $primary_size ) {
+					if ( $new_size && $primary_size && $new_size >= $primary_size ) {
 						wp_delete_file( dirname( $file ) . '/' . $new_size_meta['file'] );
 						continue;
 					}
